+++ src/index.ts
@@ -87,6 +87,7 @@ export * from './components/progress-circular';
 export * from './components/progress-linear';
 export * from './components/radio-group';
 export * from './components/rating';
+export * from './components/scroll-area';
 export * from './components/segmented-button';
 export * from './components/select';
 export * from './components/shortcut';